distributed_multi_agent/report_agent_service/requirements.txt
langchain==0.3.27
langgraph==0.6.5
langchain-openai==0.3.31
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
python-dotenv==1.1.1
